Move WorkflowEventDisplayContext into open source directory for use within workflow.execution.initiated
